chore: fresh slate and update plan
Due to the limitations of text based user input and the unreliable
rendering of colors and symbols across various terminal themes, I have
chosen to switch course and form the code from a selection of 6 numbers
(1–6) instead of a selection of 6 colors. This will avoid rendering
issues in different terminals, themes, and environments, as well as
make it more straightforward to prompt and process the user's input for
this command line program.