chore: fresh slate & redefine classes
It has become apparent that it will likely later prove useful to
distinguish between the human and the computer when it comes to which
role they are playing in the game. This should allow the code to
be more changeable following phase 1 when I look to add a computer code
breaker and human code maker. I need to embrace OOP so that I can make
those later expansions without major refactoring and rebuilding.